Bug correction about solo computation
* Feature reader forget about "distances" attribute of the graph's edges
* Can't compute spg from loaded features, some types bugs

+++ partition/graphs.py
@@ -63,8 +63,10 @@ def compute_graph_nn_2(xyz, k_nn1, k_nn2, voronoi = 0.0):
        
         graph["distances"] = graph["distances"][keep_edges]
     else:
+        # Select only "k_nn1" neighbors and distance for each point
         neighbors = neighbors[:, :k_nn1]
         distances = distances[:, :k_nn1]
+        # Source: [0,..,k_nn1, 0,..,k_nn1, .. n_ver times]
         graph["source"] = np.matlib.repmat(range(0, n_ver)
             , k_nn1, 1).flatten(order='F').astype('uint32')
         graph["target"] = np.transpose(neighbors.flatten(order='C')).astype('uint32')
@@ -152,7 +154,7 @@ def compute_sp_graph(xyz, d_max, in_component, components, labels, n_labels):
         if has_labels and label_hist:
             graph["sp_labels"][i_com, :] = sum(labels[comp,:])
         graph["sp_point_count"][i_com] = len(comp)
-        xyz_sp = np.unique(xyz[comp, :], axis=0)
+        xyz_sp = np.unique(xyz[comp.astype('uint32'), :], axis=0)
         if len(xyz_sp) == 1:
             graph["sp_centroids"][i_com] = xyz_sp
             graph["sp_length"][i_com] = 0
